+++ src/main/java/com/excilys/cdb/dto/ComputerDto.java
@@ -12,17 +12,17 @@ public class ComputerDto extends Dto {
 	@Pattern(regexp=DATE_PATTERN)
 	private String discontinued;
 	@PositiveOrZero
-	private Integer companyId;
+	private Long companyId;
 	@NotEmpty
 	private String companyName;
 	
 	private static final String DATE_PATTERN = "^(((19|2[0-9])[0-9]{2})/(0[469]|11)/(0[1-9]|[12][0-9]|30))$";
 	
-	public ComputerDto(Integer id) {
-		this(id,"",null,null,0,"None");
+	public ComputerDto(Long id) {
+		this(id,"",null,null,0L,"None");
 	}
 	
-	public ComputerDto(Integer id, String name, String intro, String disc, Integer cId, String cName) {
+	public ComputerDto(Long id, String name, String intro, String disc, Long cId, String cName) {
 		super(id);
 		this.name = name;
 		this.introduction = intro;
@@ -55,11 +55,11 @@ public class ComputerDto extends Dto {
 		this.discontinued = discontinued;
 	}
 	
-	public Integer getCompanyId() {
+	public Long getCompanyId() {
 		return companyId;
 	}
 
-	public void setCompanyId(Integer companyId) {
+	public void setCompanyId(Long companyId) {
 		this.companyId = companyId;
 	}
 
@@ -93,7 +93,7 @@ public class ComputerDto extends Dto {
 		result = 31*result + this.getName().hashCode();
 		result = 31*result + ((this.getIntroduction() == null) ? 0 : this.getIntroduction().hashCode());
 		result = 31*result + ((this.getDiscontinued() == null) ? 0 : this.getDiscontinued().hashCode());
-		result = 31*result + this.getCompanyId();
+		result = 31*result + this.getCompanyId().intValue();
 		
 		return result;
 	}
