+++ src/main/java/com/excilys/cdb/dto/Dto.java
@@ -4,17 +4,17 @@ import javax.validation.constraints.PositiveOrZero;
 
 public abstract class Dto {
 	@PositiveOrZero
-	protected Integer id;
+	protected Long id;
 	
-	public Dto(Integer id) {
+	public Dto(Long id) {
 		this.id = id;
 	}
 	
-	public Integer getId() {
+	public Long getId() {
 		return id;
 	}
 
-	public void setId(Integer id) {
+	public void setId(Long id) {
 		this.id = id;
 	}
